requirements.txt
Django
djangorestframework
Pillow
django-cors-headers
django-extensions
unidecode
django-ckeditor
django-filters
django_hosts